+++ topobank/test_challenge_redirect.py
@@ -11,5 +11,6 @@ from topobank.manager.tests.utils import UserFactory
 def test_redirect_for_challenge(client):
     user = UserFactory()
     client.force_login(user)
-    response = client.get('challenge')
-    SimpleTestCase().assertRedirects(response, settings.CHALLENGE_REDIRECT_URL)
+    response = client.get('/challenge/')
+    SimpleTestCase().assertRedirects(response, settings.CHALLENGE_REDIRECT_URL,
+                                     status_code=302, fetch_redirect_response=False)