+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - ENH: Increased step size for maximum number of iterations in 
   contact mechanics (#809)
 - ENH: Improved filenames of downloaded containers (#815)
+- ENH: Added temporary redirect to challenge URL
 - BUG: Fixed logscale labels (#771)
 - BUG: Renamed y-label of curvature distribution to probability
   density (#350)